Add criterion benchmarking infra
