+++ src/lib.rs
@@ -1,3 +1,4 @@
+pub mod fixtures;
 mod page_tracker;
 mod payload;
 mod payload_storage;
